+++ ghci-websockets.nix
@@ -1,10 +1,15 @@
-{ mkDerivation, aeson, base, foreign-store, stdenv, websockets }:
+{ mkDerivation, aeson, base, bytestring, containers, foreign-store
+, stdenv, text, wai, wai-app-static, warp, websockets
+}:
 mkDerivation {
   pname = "ghci-websockets";
   version = "0.0.1";
   src = ./.;
-  libraryHaskellDepends = [ aeson base foreign-store websockets ];
+  libraryHaskellDepends = [
+    aeson base bytestring containers foreign-store text wai
+    wai-app-static warp websockets
+  ];
   homepage = "https://github.com/j-mueller/ghci-websockets";
-  description = "Run a websocket server in GHCi";
+  description = "A websocket server that survives GHCi reloads";
   license = stdenv.lib.licenses.bsd3;
 }