+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-# 2019-07-14
+# current
+
+* Better module layout
+* Server the index.html page over HTTP
+
+# v0.0.1 (2019-07-14)
 
 * Initial release